+++ packages/globalpayments-js/types/internal/lib/installments/contracts/constants.d.ts
@@ -2,3 +2,4 @@ export declare const INSTALLMENTS_KEY = "installments";
 export declare const INSTALLMENTS_CONFIG_DEFAULT_CHANNEL = "CNP";
 export declare const INSTALLMENTS_CONFIG_DEFAULT_ENTRY_MODE = "ECOM";
 export declare const INSTALLMENTS_CONFIG_DEFAULT_CVV_INDICATOR = "PRESENT";
+export declare const INSTALLMENT_VALIDITY_KEY = "installment-validity";