+++ packages/globalpayments-js/types/lib/version.d.ts
@@ -1,2 +1,2 @@
-declare const _default: "4.1.26";
+declare const _default: "5.0.0";
 export default _default;