+++ packages/globalpayments-js/types/tools/configure.d.ts
@@ -42,8 +42,6 @@ export interface IConfiguration extends IDictionary {
             buttonLocale?: string;
             buttonStyle?: string;
             buttonType?: string;
-            countryCode?: string;
-            currencyCode?: string;
             globalPaymentsClientID?: string;
             merchantCapabilities?: string[];
             merchantIdentifier?: string;
@@ -55,7 +53,6 @@ export interface IConfiguration extends IDictionary {
             allowedCardNetworks?: string[];
             canadianDebit?: boolean;
             ctpClientId: string;
-            currencyCode?: string;
             subtotal: number;
             wrapper?: boolean;
             buttonless: boolean;
@@ -67,8 +64,6 @@ export interface IConfiguration extends IDictionary {
             buttonLocale?: string;
             buttonSizeMode?: string;
             buttonType?: string;
-            countryCode?: string;
-            currencyCode?: string;
             globalPaymentsClientID?: string;
             merchantId?: string;
             merchantName?: string;
@@ -76,7 +71,6 @@ export interface IConfiguration extends IDictionary {
         konek?: {
             buttonColor?: string;
             buttonLocale?: string;
-            countryCode?: string;
             enabled?: boolean;
             accountName: string;
             channel: string;